live stream corrections
